Do not build this over and over

+++ font-preview/gen.sh
@@ -2,8 +2,8 @@
 
 set -e
 
-cd $(dirname $(basename $0));
-
+cd $(dirname $0);
+echo "Working in `pwd`";
 mkdir -p ../root/static/images/font-preview
 
 for font in 'CMU Serif'            \
@@ -27,17 +27,21 @@ for font in 'CMU Serif'            \
             'DejaVu Sans Mono'     \
             'TeX Gyre Cursor'; do
     rm -f font-preview.{tex,log,pdf}
+    pdf=$(echo $font | sed -e 's/ /-/g').pdf
+    if [ ! -f "../root/static/images/font-preview/$pdf" ]; then
+
     muse-compile.pl --extra papersize=a5 --extra division=15 \
         --extra sitename="$font" \
         --extra nocoverpage=1 \
         --extra fontsize=11 --extra mainfont="$font" \
         --pdf font-preview.muse >/dev/null 2>&1
-    pdf=$(echo $font | sed -e 's/ /-/g').pdf
+
     mv font-preview.pdf $pdf
     png=$(basename $pdf .pdf).png
     convert -density 150 -trim -quality 100 -sharpen 0x1.0 $pdf[1] $png
     echo "created $pdf and $png"
     mv $pdf $png ../root/static/images/font-preview
     rm -f font-preview.{aux,tex,log}
+    fi
 done
